+++ apps/smtp/scripts/deploy.ts
@@ -7,9 +7,3 @@ import packageJson from "../package.json";
 exportSentryReleaseEnvironmentVariable(packageJson.version);
 
 execSync("pnpm run build", { stdio: "inherit" });
-/*
- * execSync("pnpm --silent run dump-templates | pnpm --silent run audit-helpers", {
- *   stdio: "inherit",
- *   shell: "/bin/sh",
- * });
- */